+++ controller/machine/scope_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers //nolint:testpackage
+package machine //nolint:testpackage
 
 import (
 	"testing"
